Migrate half of routes/shared.ts importers to the split files (1/2)
Mechanical follow-up to the routes/shared.ts split: repoints these
importers' import paths at validation.ts/viewHelpers.ts/errorHandling.ts/
uploadMiddleware.ts directly instead of the temporary shared.ts barrel.
No logic changes. The barrel stays in place until the remaining
importers are migrated in a follow-up PR, which also deletes it.
