+++ src/web/routes/adminUserValidation.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { createLogger } from '../../shared/logger';
 import { findUser, getMemberAccessLevel, AccessLevel } from '../../db';
-import { trimField, normalizeDiscordId } from './shared';
+import { trimField, normalizeDiscordId } from './validation';
 import { getSessionUser } from '../session';
 import { normalizeTwitchChannelName } from '../../twitch/twitchChannelName';
 import { isLockWaitTimeoutDbError } from './adminUserMutations';
