+++ src/web/routes/alertsAdminMutations.ts
@@ -3,7 +3,9 @@ import { Router } from 'express';
 import { csrfProtection } from '../csrf';
 import { requireAuth } from '../middleware';
 import { saveAlertConfig, getAlertConfig, ALERT_TEXT_ANIMATIONS } from '../../db';
-import { logAndRedirectError, requireStreamer, parseCheckboxField } from './shared';
+import { parseCheckboxField } from './validation';
+import { requireStreamer } from './viewHelpers';
+import { logAndRedirectError } from './errorHandling';
 import { pushAlertEvent } from './alertsOverlaySource';
 import { NOT_A_STREAMER_REDIRECT, parseEventType, parseEnumField } from './alertsShared';
 import { reloadEventSubSubscriptions } from '../../twitch/eventsub/twitchEventSub';
